Adding Combinations of keys per command
I didn't want to have to rely on something like autohotkey just to press
a couple of keys in sequence from a single command. Added up to 7
additional keypresses per command. Also changed the X close button to
actually close the window and the minimize button to minimize to system
tray. (always been a peeve of mine when close doesn't close)